can edit and delete schedule, structuring data access source code

+++ HomeBase/HomeBase/Player.swift
@@ -12,10 +12,10 @@ class Player: NSObject {
     
     var playerID: Int64
     var name: String
-    var backNumber: Int64
+    var backNumber: Int
     var position: String
     
-    init(id: Int64 = 0, name: String, backNumber: Int64, position: String) {
+    init(id: Int64 = 0, name: String, backNumber: Int, position: String) {
         self.playerID = id
         self.name = name
         self.backNumber = backNumber
